env and env dependencies progress

+++ src/burn/agent/base/mod.rs
@@ -14,4 +14,4 @@ pub use model::Model;
 pub use snapshot::Snapshot;
 pub use state::State;
 
-pub type ElemType = f32;
+pub type ElemType = f64;
